requirements.txt
asgiref==3.5.2
autopep8==1.6.0
click==8.1.3
colorama==0.4.4
Django==3.0.3
django-tempus-dominus==5.1.2.17
django-widget-tweaks==1.4.12
djlint==1.1.1
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==4.11.4
pathspec==0.9.0
pycodestyle==2.8.0
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
zipp==3.8.0
